+++ cmake/GoogleCloudCppFeatures.cmake
@@ -25,7 +25,7 @@ set(GOOGLE_CLOUD_CPP_LEGACY_FEATURES
 
 set(GOOGLE_CLOUD_CPP_EXPERIMENTAL_LIBRARIES # cmake-format: sorted
     # This is WIP, it needs a number of hand-crafted APIs.
-    "pubsublite")
+    "pubsublite" "sql")
 
 set(GOOGLE_CLOUD_CPP_TRANSITION_LIBRARIES # cmake-format: sorted
     # Transitioned circa 2023-02-22